requirements.txt
streamlit
textblob
pyspellchecker